add recursive recipes
Recipes may now contain other recipes as ingredients. This makes it
easier to define certain kinds of foods as recipes (e.g. bread, pasta,
etc.) and still be able to include them in other recipes later on.
This also paves the way for 'menus', which are just collections of
recipes.